requirements.txt
huggingface_hub
python-dotenv
Pillow
